API - Added JWT Token and changed Model of User

+++ api/controllers/userController.js
@@ -1,9 +1,12 @@
 const UserModel = require('../models/User');
-
 const userController = {};
 
 userController.getAllUsers = async (req, res) => {
     try {
+        if (!req.user.isAdmin) {
+            return res.status(403).json({ message: "Unauthorized: Only administrators can perform this action" });
+        }
+        
         const users = await UserModel.find();
         res.json(users);
     } catch (error) {
@@ -22,10 +25,19 @@ userController.getUserById = async (req, res) => {
 
 userController.createUser = async (req, res) => {
     try {
-        const userData = { ...req.body };
-        if (!userData.lastUpdated) {
-            userData.lastUpdated = Date.now();
+        if (!req.user.isAdmin) {
+            return res.status(403).json({ message: "Unauthorized: Only administrators can perform this action" });
         }
+
+        const { email } = req.body;
+        const existingUser = await UserModel.findOne({ email });
+        if (existingUser) {
+            return res.status(400).json({ message: "User with this email already exists" });
+        }
+
+        const userData = { ...req.body };
+        userData.lastUpdated = Date.now();
+
         const user = await UserModel.create(userData);
         res.json(user);
     } catch (error) {
@@ -35,10 +47,13 @@ userController.createUser = async (req, res) => {
 
 userController.updateUser = async (req, res) => {
     try {
-        const userData = { ...req.body };
-        if (!userData.lastUpdated) {
-            userData.lastUpdated = Date.now();
+        if (!req.user.isAdmin) {
+            return res.status(403).json({ message: "Unauthorized: Only administrators can perform this action" });
         }
+
+        const userData = { ...req.body };
+        userData.lastUpdated = Date.now();
+
         const user = await UserModel.findByIdAndUpdate(req.params.id, userData, { new: true });
         res.json(user);
     } catch (error) {
@@ -48,6 +63,10 @@ userController.updateUser = async (req, res) => {
 
 userController.deleteUser = async (req, res) => {
     try {
+        if (!req.user.isAdmin) {
+            return res.status(403).json({ message: "Unauthorized: Only administrators can perform this action" });
+        }
+
         const deletedUser = await UserModel.findByIdAndDelete(req.params.id);
         res.json(deletedUser);
     } catch (error) {
